// src/components/FileItem/components/ExtendedVariant/stylesBuilder.d.ts
import { ExtendedVariantConfig } from './ExtendedVariant.styles';
import { ProgressBarConfig } from '../ProgressBar';
import { BaseProps } from '../../../../types/BaseProps';
type ExtendedVariantStylesBuilder = {
    container: BaseProps;
    thumbnailWrapper: BaseProps;
    fileDetails: BaseProps;
    topDetails: BaseProps;
    bottomDetails: BaseProps;
    fileThumbnail: BaseProps;
    photoThumbnail: BaseProps;
    fileDescription: BaseProps;
    fileName: BaseProps;
    fileSize: BaseProps;
    timeLeft: BaseProps;
    dot: BaseProps;
    uploadedPercentage: BaseProps;
    uploadingContent: BaseProps;
    replaceableContent: BaseProps;
    alertContent: BaseProps;
    closeIconButton: BaseProps;
    alert: BaseProps;
    alertText: BaseProps;
    alertIcon: BaseProps;
    fileSizeAlert: BaseProps;
    progressBar: ProgressBarConfig;
};
export declare const stylesBuilder: (custom?: ExtendedVariantConfig) => ExtendedVariantStylesBuilder;
export {};